pyproject.toml
[project]
name = "mcp-server-template-python"
version = "0.1.0"
description = "A starter template for building Model Context Protocol (MCP) servers in Python"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.9.1",
    "fastapi>=0.105.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.2",
    "loguru>=0.7.2",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
mcp-server = "src.mcp_server.cli:main"

